Moves the ICaseType constant to the consts lib
