+++ apps/judicial-system/web/src/routes/Prosecutor/InvestigationRequest/Defendant/DefendantForm.tsx
@@ -7,7 +7,8 @@ import {
   FormContentContainer,
   FormFooter,
 } from '@island.is/judicial-system-web/src/shared-components'
-import { Case, CaseType, ICaseTypes } from '@island.is/judicial-system/types'
+import { ICaseTypes } from '@island.is/judicial-system/consts'
+import { Case, CaseType } from '@island.is/judicial-system/types'
 import {
   removeTabsValidateAndSet,
   setAndSendToServer,
